functests/requirements.txt
mockserver-client
pika
pymysql